Добавляет actionConvert в DataController
